refactor: split new landing page into components
