+++ web/src/recoil/api/sdl.ts
@@ -44,7 +44,7 @@ export const fetchLandingPageMetadata: QueryFunction<any, string> = async ({quer
   const octokit = new Octokit({});
 
   const repository = await octokit.request('GET /repos/{owner}/{repo}/contents/{path}/metadata.json', {
-    owner: 'kaushik-rishi',
+    owner: 'akash-network',
     repo: 'deploy-templates',
     path: 'landing',
   });
